acpi: ged: Separate GED code from hardware reduced requirement
Here we move the GED specific code to its own files. This way, it
is made easier to separate the hotplug handled through GED from the
ACPI hardware reduced specification.
